pyproject.toml
[tool.poetry]
name = "tarot-cards"
version = "0.1.0"
description = ""
authors = ["Plaintextnerds <tim@plaintextnerds.com>"]
readme = "README.md"
packages = [{include = "plugin"}]

[tool.poetry.dependencies]
python = "^3.9,<3.13"
cloudquery-plugin-sdk = "^0.1.12"
pydantic = "^2.4"
pyarrow = "^14.0.2"
requests = "^2.32.2"

[tool.poetry.scripts]
main = "main:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

